Implement feedback from draft review, extract annotation element to its own file
